fixed usersController
